Create first testing Web API for Rooms
