+++ intraweb/src/intraweb/Startup.cs
@@ -16,6 +16,7 @@ namespace intraweb
 {
     public class Startup
     {
+
         public Startup(IHostingEnvironment env)
         {
             // Set up configuration sources.
@@ -53,6 +54,8 @@ namespace intraweb
             // Add application services.
             services.AddTransient<IEmailSender, AuthMessageSender>();
             services.AddTransient<ISmsSender, AuthMessageSender>();
+
+            services.AddInstance<IRoomRepository>(new Models.Dummies.RoomDummyRepository()); //Testovacia implementacia
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
