+++ intraweb/src/intraweb/appsettings.json
@@ -5,7 +5,7 @@
     }
   },
   "Logging": {
-    "IncludeScopes": false,
+    "IncludeScopes": false,    
     "LogLevel": {
       "Default": "Verbose",
       "System": "Information",
